pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-readme-improver"
version = "0.1.0"
description = "CLI tool to improve README files using OpenAI"
authors = [{name = "OpenAI", email = "example@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai",
    "python-dotenv",
    "markdown2"
]

[tool.setuptools]
packages = ["readme_improver"]
py-modules = [
    "main",
]

[project.scripts]
readme-improver = "main:main"
